Rholang: Normalizer: Add DeBruijnIndexMap but don't use it.
Using Debruijn indices will make matching simpler. Add the map, but don't use it yet.
