+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/DebruijnIndexMap.scala
@@ -1,0 +1,66 @@
+package coop.rchain.rholang.interpreter
+
+// This is an index map. Note that the internal environment is the same as the
+// level map, but we calculate the correct index on get.
+class DebruijnIndexMap[T](val next: Int, val env: Map[String, (Int, T)]) {
+  def this() = this(0, Map[String, (Int, T)]())
+
+  def newBinding(binding: (String, T)): DebruijnIndexMap[T] =
+    binding match {
+      case (varName, sort) =>
+        DebruijnIndexMap[T](next + 1, env + (varName -> ((next, sort))))
+    }
+
+  // Returns the new map
+  def newBindings(bindings: List[(String, T)]): DebruijnIndexMap[T] =
+    (this /: bindings)((map, binding) => map.newBinding(binding))
+
+  // Returns the new map, and a list of the shadowed variables
+  // Takes a **Level** map, because we use that to track the Free Variables.
+  def absorbFree(binders: DebruijnLevelMap[T]): (DebruijnIndexMap[T], List[String]) = {
+    val finalNext  = next + binders.next
+    val adjustNext = next
+    binders.env.foldLeft((this, List[String]())) {
+      case ((db: DebruijnIndexMap[T], shadowed: List[String]),
+            (k: String, (level: Int, varType: T @unchecked))) => {
+        val shadowedNew = if (db.env.contains(k)) k :: shadowed else shadowed
+        (DebruijnIndexMap(finalNext, db.env + (k -> ((level + adjustNext, varType)))), shadowedNew)
+      }
+    }
+  }
+
+  def getBinding(varName: String): Option[T] =
+    for (pair <- env.get(varName)) yield pair._2
+  def getLevel(varName: String): Option[Int] =
+    for (pair <- env.get(varName)) yield pair._1
+  def get(varName: String): Option[(Int, T)] =
+    env
+      .get(varName)
+      .map({
+        case (level, t) => (next - level - 1, t)
+      })
+  def isEmpty() = next == 0
+
+  def count: Int = next
+
+  override def equals(that: Any): Boolean =
+    that match {
+      case that: DebruijnLevelMap[T] =>
+        next == that.next &&
+          env == that.env
+      case _ => false
+    }
+
+  override def hashCode(): Int =
+    (next.hashCode() * 37 + env.hashCode)
+}
+
+object DebruijnIndexMap {
+  def apply[T](next: Int, env: Map[String, (Int, T)]): DebruijnIndexMap[T] =
+    new DebruijnIndexMap(next, env)
+
+  def apply[T](): DebruijnIndexMap[T] = new DebruijnIndexMap[T]()
+
+  def unapply[T](db: DebruijnIndexMap[T]): Option[(Int, Map[String, (Int, T)])] =
+    Some((db.next, db.env))
+}
